Updates sequencer to handle multiple txs
